requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
psycopg2==2.9.10
Pillow==11.0.0
drf-spectacular==0.27.2